changes: tag disc number implementation complete toDo: use disc number for sorting use album artist for file organisation on disc

+++ jajuk/src/main/java/org/jajuk/services/tags/Tag.java
@@ -419,7 +419,7 @@ public class Tag {
   /**
    * @param discnumber
    */
-  public void setDiscNumber(int discnumber) throws JajukException {
+  public void setDiscNumber(long discnumber) throws JajukException {
     try {
       tagImpl.setDiscNumber(discnumber);
     } catch (Exception e) {
